web/requirements.txt
flask 
azure-cosmos

